requirement.txt
aiosmtplib==1.1.6
anyio==3.6.1
blinker==1.5
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-mail==1.1.4
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
pyasn1==0.4.8
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
